Fix for date format from epoch unix s to ms

+++ UnityPerformanceBenchmarkReporter/Entities/Data.cs
@@ -38,7 +38,7 @@ namespace UnityPerformanceBenchmarkReporter.Entities.New
     public class Run
     {
         public string TestSuite;
-        public int Date;
+        public long Date;
         public Player Player;
         public Hardware Hardware;
         public Editor Editor;
